Make Projects Expand and Create Center Form
This commit is mainly focused on adding functionality to the Projects
displayed on the sidebar. Some things that are included are the
expanding and retracting buttons added as well as the show more and
show less buttons. Both of these were made so that they don't have to
be fed a link to a specific element and instead navigate to it on their
own which will become helpful later.

The center form has also been created with some relatively simplistic
CSS applied. This will be changed more in the future.