Finish JS Additions and Make UI Pretty
The main additions this commit adds are teh finished Javascript sections
and making the user interface more appealing.

Many things were added to the JS, notably adding in the user storage and
making the buttons more 'dummy proof'. There are some more things that
need to be added like requiring the project and task names to be unique
and filled out.

The Color choices are all shades of a mint green. While not exceptionally
appealing it gets the job done which is all I need. Special care was taken
to make padding and margins line up properly.